pyproject.toml
[project]
name = "costbot"
version = "0.1.0"
requires-python = ">=3.13"

dependencies = [
    "aiohttp>=3.13.3",
    "asyncio>=4.0.0",
    "fastapi>=0.128.8",
    "python-dotenv>=1.2.1",
    "tzdata>=2025.3",
    "uvicorn>=0.39.0",
]

[dependency-groups]
dev = [
    "black>=25.11.0",
    "isort>=6.1.0",
]
